config: separate credentials.lua into two modules
This patch moves part of credentials.lua to a new module. The new module
is not complete and will be properly introduces in the next commit.

Part of #9643

NO_DOC=will be added in the next commit
NO_TEST=will be added in the next commit
NO_CHANGELOG=will be added in the next commit
